+++ grunt/config/sass.js
@@ -11,6 +11,7 @@ module.exports = {
 	example: {
 		files: {
 			"examples/browserified/style.css": "examples/browserified/style.scss",
+			"examples/relevant-words-example/style.css": "examples/relevant-words-example/style.scss",
 		},
 	},
 };